BonaAkademia/projekt-main/app/requirements.txt
asgiref==3.5.2
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
django-rest-swagger==2.2.0
importlib-metadata==4.12.0
Markdown==3.4.1
psycopg2-binary==2.9.1
pytz==2022.2.1
sqlparse==0.4.2
zipp==3.8.1
django-environ==0.4.*
django-cors-headers==3.11.*
